pyproject.toml
[project]
name = "ink-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all]>=0.115.8",
    "pillow>=11.1.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "rpi-gpio>=0.7.1",
    "spidev>=3.6",
]
